requirements.txt
# Core numeric / utilities
numpy>=1.26.0

# Retrieval stack
sentence-transformers>=2.6.1
torch>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Google Gemini SDK + deps (v1beta compatible)
google-generativeai>=0.7.0
google-ai-generativelanguage>=0.6.5
grpcio>=1.60.0
protobuf>=4.25.3

# Env & PDF extraction
python-dotenv>=1.0.1
pymupdf4llm>=0.0.17
pymupdf>=1.23.8
